refactor(server): align subscriptions domain structure
