+++ apps/server/src/domain/subscriptions/index.ts
@@ -3,12 +3,12 @@ export * from "./models";
 export * from "./repository/subscription-command.repository";
 export * from "./repository/subscription-query.repository";
 export * from "./repository/subscription.repository.types";
-export * from "./services/activate-subscription.service";
-export * from "./services/create-subscription.service";
-export * from "./services/subscription-command.live";
-export * from "./services/subscription-command.service";
-export * from "./services/subscription-flows.shared";
-export * from "./services/subscription-query.live";
-export * from "./services/subscription-query.service";
+export * from "./services/commands/activate-subscription.service";
+export * from "./services/commands/create-subscription.service";
+export * from "./services/commands/subscription-command.live";
+export * from "./services/commands/subscription-command.service";
+export * from "./services/commands/use-subscription-once.service";
+export * from "./services/queries/subscription-query.live";
+export * from "./services/queries/subscription-query.service";
+export * from "./services/shared/subscription-flow.shared";
 export * from "./services/subscription.service.types";
-export * from "./services/use-subscription-once.service";
