+++ apps/server/src/domain/subscriptions/services/shared/subscription-flow.shared.ts
@@ -1,9 +1,9 @@
 import { env } from "@/config/env";
 
-import type { ActiveSubscriptionExists, SubscriptionExpired, SubscriptionNotFound, SubscriptionNotPending, SubscriptionNotUsable, SubscriptionPendingOrActiveExists, SubscriptionUsageExceeded } from "../domain-errors";
+import type { ActiveSubscriptionExists, SubscriptionExpired, SubscriptionNotFound, SubscriptionNotPending, SubscriptionNotUsable, SubscriptionPendingOrActiveExists, SubscriptionUsageExceeded } from "../../domain-errors";
 
 /**
- * Tập lỗi mà flow dùng một lượt subscription có thể trả ra.
+ * Tap loi ma flow dung mot luot subscription co the tra ra.
  */
 export type UseSubscriptionFailure
   = | SubscriptionNotFound
@@ -13,15 +13,15 @@ export type UseSubscriptionFailure
     | ActiveSubscriptionExists;
 
 /**
- * Tập lỗi của flow tạo subscription mới.
+ * Tap loi cua flow tao subscription moi.
  */
 export type CreateSubscriptionFailure
   = | SubscriptionPendingOrActiveExists
-    | import("../../wallets/domain-errors").InsufficientWalletBalance
-    | import("../../wallets/domain-errors").WalletNotFound;
+    | import("../../../wallets/domain-errors").InsufficientWalletBalance
+    | import("../../../wallets/domain-errors").WalletNotFound;
 
 /**
- * Tập lỗi của flow kích hoạt subscription.
+ * Tap loi cua flow kich hoat subscription.
  */
 export type ActivateSubscriptionFailure
   = | SubscriptionNotFound
@@ -32,14 +32,18 @@ const EXPIRE_AFTER_MS = env.EXPIRE_AFTER_DAYS * 24 * 60 * 60 * 1000;
 const AUTO_ACTIVATE_MS = env.AUTO_ACTIVATE_IN_DAYS * 24 * 60 * 60 * 1000;
 
 /**
- * Tính mốc hết hạn từ một thời điểm bắt đầu.
+ * Tinh moc het han tu mot thoi diem bat dau.
+ *
+ * @param now Moc kich hoat subscription.
  */
 export function computeExpiresAt(now: Date): Date {
   return new Date(now.getTime() + EXPIRE_AFTER_MS);
 }
 
 /**
- * Tính mốc auto-activate cho job nền sau khi user mua gói.
+ * Tinh moc auto-activate cho job nen sau khi user mua goi.
+ *
+ * @param now Moc tao subscription pending.
  */
 export function computeAutoActivateAt(now: Date): Date {
   return new Date(now.getTime() + AUTO_ACTIVATE_MS);
